test: fold auto-mock tests into mock-module.test.ts
Move the auto-mock + requireMock coverage from a standalone
auto-mock.test.ts into the existing mock-module.test.ts file, under
a clearly-marked section. Matches the repo convention (CLAUDE.md:
"add your test to the existing test file for the code you're
changing") — mock-module.test.ts already covers factory-based
`mock.module`, so the factory-less auto-mock path belongs with it.

Fixtures (auto-mock-fixture*.ts) remain their own files — keeping
multiple fixture sources inline would bloat the test file.
